removed misplaced TextureMatrix from SMaterial.
the correct function is driver->setTransform ( TextureXX )
(beside it tooks ca. 384 byte per SMaterial..)

how would you do this for example?. You have 2000 walls with the same moving banner, but the banner is always on a different position. would you have 2000 different Materials?. No, just 1 Material and 2000 different transformations. or ?

really;-) it makes no sense to store a Texturematrix in the Material.


git-svn-id: svn://svn.code.sf.net/p/irrlicht/code@457 dfc29bdd-3216-0410-991c-e03cc46cb475